requirements.txt
requests>=2.28.1
python-dotenv>=0.21.0
colorama>=0.4.6
